+++ matvecops.hpp
@@ -4,21 +4,21 @@
 #include <vector>
 
 // take sum of vector elements
-double sum(std::vector<double>& v);
+long double sum(std::vector<long double>& v);
 
 // add or subtract two same-length vectors
-std::vector<double> add(
-    std::vector<double>& v1,
-    std::vector<double>& v2,
+std::vector<long double> add(
+    std::vector<long double>& v1,
+    std::vector<long double>& v2,
     bool subtract);
 
 // scalar multiplication between vector and constant
-std::vector<double> scale(std::vector<double>& v, double c);
+std::vector<long double> scale(std::vector<long double>& v, long double c);
 
 // dot product between two vectors
-double dot(std::vector<double>& v1, std::vector<double>& v2);
+long double dot(std::vector<long double>& v1, std::vector<long double>& v2);
 
 // print vector to cout, useful for debugging
-void print_vec(std::vector<double>& v);
+void print_vec(std::vector<long double>& v);
 
 #endif
